Remove course icons from images directory
The images directory hold static assets that are served. The icons
are associated with the course and is loaded by a loop in the
courses#index view so they can't be individually linked. I need to
setup that the course icon is stored in the database so it can be
loaded with the course.