Reworked CUBE CSS to remove utility generation
The utility generation vs. manually writing out the utilities was just
too complex and burdensome. With so many utilities written out I doubt
it well ever be needed to add utilities, and if so I doubt it will be
often or a lot. Leave it simple for now and if generation is needed in
the future it can be added.