Change the root path to the main marketing page
When users type in korsarium.com they should be taken to the main
marketing page where they will be prompted to sign up or login in the
nav.

Courses#index can no longer be the root path because it requires
authentication so when users would try to go to korsarium.com they would
be automatically redirected to the sessions#new view which would prompt
them to login which omits a step for brand new users.